event page regactoring
